Move `tapConsole(...)` to shareable fixtures
