+++ flogger/flogger-api-tests/build.gradle.kts
@@ -31,4 +31,5 @@ plugins {
 dependencies {
     testImplementation(project(":flogger-api"))
     testImplementation(project(":flogger-testing"))
+    testImplementation(project(":fixtures"))
 }
