+++ logging/build.gradle.kts
@@ -87,6 +87,7 @@ kotlin {
         }
         val jvmTest by getting {
             dependencies {
+                implementation(project(":fixtures"))
                 implementation(project(":flogger-system-backend"))
                 implementation(Spine.testlib) {
                     exclude(group = "com.google.flogger")
